- fixed barrack to not stop requesting weapons - reworked creation of soldiers

+++ jsettlers.logic.buildings/jsettlers/logic/stack/IRequestsStackGrid.java
@@ -11,6 +11,6 @@ public interface IRequestsStackGrid {
 
 	void popMaterial(ISPosition2D position, EMaterialType materialType);
 
-	int getStackSize(ISPosition2D position, EMaterialType materialType);
+	byte getStackSize(ISPosition2D position, EMaterialType materialType);
 
 }
