+++ jsettlers.logic.buildings/jsettlers/logic/stack/RequestStack.java
@@ -45,8 +45,8 @@ public class RequestStack {
 		return materialType;
 	}
 
-	public int getStackSize() {
-	    return grid.getStackSize(position, materialType);
-    }
+	public byte getStackSize() {
+		return grid.getStackSize(position, materialType);
+	}
 
 }
